Share `no-constant-condition`'s `isConstant`
Here we extract `isConstant` into ast-utils and share it between
`no-constant-condition` and `no-constant-binary-expression`.
