+++ lib/rules/no-constant-binary-expression.js
@@ -5,83 +5,13 @@
 
 "use strict";
 
-const { isNullOrUndefined } = require("./utils/ast-utils");
+const { isNullOrUndefined, isConstant } = require("./utils/ast-utils");
 
 const NUMERIC_OR_STRING_BINARY_OPERATORS = new Set(["+", "-", "*", "/", "%", "|", "^", "&"]);
 
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
-
-/**
- * Test if an AST node has a statically knowable constant truthiness. Meaning,
- * it will always coerce to either `true` or `false` when cast directly to
- * boolean.
- * @param {ASTNode} node The AST node being tested.
- * @returns {boolean} Does `node` have constant truthiness?
- */
-function hasConstantTruthiness(node) {
-    switch (node.type) {
-        case "ObjectExpression": // Objects are always truthy
-        case "ArrayExpression": // Arrays are always truthy
-        case "ArrowFunctionExpression": // Functions are always truthy
-        case "FunctionExpression": // Functions are always truthy
-        case "ClassExpression": // Classes are always truthy
-        case "NewExpression": // Objects are always truthy
-        case "Literal": // Truthy, or falsy, literals never change
-            return true;
-        case "CallExpression": {
-            if (node.callee.type === "Identifier" && node.callee.name === "Boolean") {
-                return node.arguments.length === 0 || hasConstantTruthiness(node.arguments[0]);
-            }
-            return false;
-        }
-        case "JSXElement": // ESLint has a policy of not assuming any specific JSX behavior.
-        case "JSXFragment":
-            return false;
-        case "AssignmentExpression":
-            if (node.operator !== "=") {
-                return false; // We won't go so far as to try to evaluate += etc.
-            }
-            return hasConstantTruthiness(node.right);
-        case "TemplateLiteral":
-
-            /*
-             * Possible future direction if needed: If all quasis are empty, we
-             * could look at node.expressions and try to determine if they are
-             * static truthinesss.
-             */
-            return node.quasis.some(quasi => quasi.value.cooked.length);
-        case "UnaryExpression":
-            if (node.operator === "void" || // Always returns `undefined`
-                node.operator === "typeof" // All type strings are truthy
-            ) {
-                return true;
-            }
-            if (node.operator === "!") {
-                return hasConstantTruthiness(node.argument);
-            }
-
-            /*
-             * We won't try to reason about +, -, ~, or delete
-             * Possible future direction if needed: In theory, for the
-             * mathematical operators, we could look at the argument and try to
-             * determine if it coerces to a constant numeric value.
-             */
-            return false;
-        case "SequenceExpression": {
-            const last = node.expressions[node.expressions.length - 1];
-
-            return hasConstantTruthiness(last);
-        }
-        case "Identifier": {
-            return node.name === "undefined";
-        }
-
-        default:
-            return false;
-    }
-}
 
 /**
  * Test if an AST node has a statically knowable constant nullishness. Meaning,
@@ -169,18 +99,19 @@ function hasConstantNullishness(node) {
  * 1. Literal booleans (`true` or `false`)
  * 2. Unary `!` expressions with a constant value
  * 3. Constant booleans created via the `Boolean` global function
+ * @param {Scope} scope The scope in which the node was found.
  * @param {ASTNode} node The node to test
  * @returns {boolean} Is `node` guaranteed to be a boolean?
  */
-function isStaticBoolean(node) {
+function isStaticBoolean(scope, node) {
     switch (node.type) {
         case "Literal":
             return typeof node.value === "boolean";
         case "CallExpression":
             return node.callee.type === "Identifier" && node.callee.name === "Boolean" &&
-         (node.arguments.length === 0 || hasConstantTruthiness(node.arguments[0]));
+         (node.arguments.length === 0 || isConstant(scope, node.arguments[0]), true);
         case "UnaryExpression":
-            return node.operator === "!" && hasConstantTruthiness(node.argument);
+            return node.operator === "!" && isConstant(scope, node.argument, true);
         default:
             return false;
     }
@@ -195,10 +126,11 @@ function isStaticBoolean(node) {
  *
  * Javascript `==` operator works by converting the boolean to `1` (true) or
  * `+0` (false) and then checks the values `==` equality to that number.
- * @param {ASTNode} node The node to test
+ * @param {Scope} scope The scope in which node was found.
+ * @param {ASTNode} node The node to test.
  * @returns {boolean} Will `node` always coerce to the same boolean value?
  */
-function hasConstantLooseBooleanComparison(node) {
+function hasConstantLooseBooleanComparison(scope, node) {
     switch (node.type) {
         case "ObjectExpression":
         case "ClassExpression":
@@ -241,7 +173,7 @@ function hasConstantLooseBooleanComparison(node) {
                 return true;
             }
             if (node.operator === "!") {
-                return hasConstantTruthiness(node.argument);
+                return isConstant(scope, node.argument, true);
             }
 
             /*
@@ -255,7 +187,7 @@ function hasConstantLooseBooleanComparison(node) {
             return false;
         case "CallExpression": {
             if (node.callee.type === "Identifier" && node.callee.name === "Boolean") {
-                return node.arguments.length === 0 || hasConstantTruthiness(node.arguments[0]);
+                return node.arguments.length === 0 || isConstant(scope, node.arguments[0], true);
             }
             return false;
         }
@@ -273,7 +205,7 @@ function hasConstantLooseBooleanComparison(node) {
             return node.expressions.length === 0;
         case "AssignmentExpression":
             if (node.operator === "=") {
-                return hasConstantLooseBooleanComparison(node.right);
+                return hasConstantLooseBooleanComparison(scope, node.right);
             }
 
             /*
@@ -288,7 +220,7 @@ function hasConstantLooseBooleanComparison(node) {
         case "SequenceExpression": {
             const last = node.expressions[node.expressions.length - 1];
 
-            return hasConstantLooseBooleanComparison(last);
+            return hasConstantLooseBooleanComparison(scope, last);
         }
         case "JSXElement": // ESLint has a policy of not assuming any specific JSX behavior.
         case "JSXFragment":
@@ -303,11 +235,12 @@ function hasConstantLooseBooleanComparison(node) {
  * Test if an AST node will always give the same result when _strictly_ compared
  * to a bolean value. This can happen if the expression can never be boolean, or
  * if it is always the same boolean value.
+ * @param {Scope} scope The scope in which the node was found.
  * @param {ASTNode} node The node to test
  * @returns {boolean} Will `node` always give the same result when compared to a
  * static boolean value?
  */
-function hasConstantStrictBooleanComparison(node) {
+function hasConstantStrictBooleanComparison(scope, node) {
     switch (node.type) {
         case "ObjectExpression": // Objects are not booleans
         case "ArrayExpression": // Arrays are not booleans
@@ -326,7 +259,7 @@ function hasConstantStrictBooleanComparison(node) {
                 return false;
             }
             if (node.operator === "!") {
-                return hasConstantTruthiness(node.argument);
+                return isConstant(scope, node.argument, true);
             }
 
             /*
@@ -338,13 +271,13 @@ function hasConstantStrictBooleanComparison(node) {
         case "SequenceExpression": {
             const last = node.expressions[node.expressions.length - 1];
 
-            return hasConstantStrictBooleanComparison(last);
+            return hasConstantStrictBooleanComparison(scope, last);
         }
         case "Identifier":
             return node.name === "undefined";
         case "AssignmentExpression":
             if (node.operator === "=") {
-                return hasConstantStrictBooleanComparison(node.right);
+                return hasConstantStrictBooleanComparison(scope, node.right);
             }
 
             /*
@@ -409,16 +342,17 @@ function isAlwaysNew(node) {
 
 /**
  * Checks if one operand will cause the result to be constant.
+ * @param {Scope} scope Scope in which the expression was found.
  * @param {ASTNode} a One side of the expression
  * @param {ASTNode} b The other side of the expression
  * @param {string} operator The binary expression operator
  * @returns {ASTNode | null} The node which will cause the expression to have a constant result.
  */
-function findBinaryExpressionConstantOperand(a, b, operator) {
+function findBinaryExpressionConstantOperand(scope, a, b, operator) {
     if (operator === "==" || operator === "!=") {
         if (
             (isNullOrUndefined(a) && hasConstantNullishness(b)) ||
-            (isStaticBoolean(a) && hasConstantLooseBooleanComparison(b)) ||
+            (isStaticBoolean(scope, a) && hasConstantLooseBooleanComparison(scope, b)) ||
 
             /*
              * If both sides are "new", then both sides are objects and
@@ -432,7 +366,7 @@ function findBinaryExpressionConstantOperand(a, b, operator) {
     } else if (operator === "===" || operator === "!==") {
         if (
             (isNullOrUndefined(a) && hasConstantNullishness(b)) ||
-            (isStaticBoolean(a) && hasConstantStrictBooleanComparison(b))
+            (isStaticBoolean(scope, a) && hasConstantStrictBooleanComparison(scope, b))
         ) {
             return b;
         }
@@ -467,17 +401,17 @@ module.exports = {
             LogicalExpression(node) {
                 const { operator, left } = node;
 
-                if ((operator === "&&" || operator === "||") && hasConstantTruthiness(left)) {
+                if ((operator === "&&" || operator === "||") && isConstant(context.getScope(), left, true)) {
                     context.report({ node: left, messageId: "constantShortCircuit", data: { property: "truthiness", operator } });
                 } else if (operator === "??" && hasConstantNullishness(left)) {
                     context.report({ node: left, messageId: "constantShortCircuit", data: { property: "nullishness", operator } });
                 }
             },
             BinaryExpression(node) {
-
+                const scope = context.getScope();
                 const { right, left, operator } = node;
-                const rightConstantOperand = findBinaryExpressionConstantOperand(left, right, operator);
-                const leftConstantOperand = findBinaryExpressionConstantOperand(right, left, operator);
+                const rightConstantOperand = findBinaryExpressionConstantOperand(scope, left, right, operator);
+                const leftConstantOperand = findBinaryExpressionConstantOperand(scope, right, left, operator);
 
                 if (rightConstantOperand) {
                     context.report({ node: rightConstantOperand, messageId: "constantBinaryOperand", data: { operator, otherSide: "left" } });
