+++ lib/rules/no-constant-condition.js
@@ -5,6 +5,8 @@
 
 "use strict";
 
+const { isConstant } = require("./utils/ast-utils");
+
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
@@ -54,208 +56,13 @@ module.exports = {
         //--------------------------------------------------------------------------
 
         /**
-         * Returns literal's value converted to the Boolean type
-         * @param {ASTNode} node any `Literal` node
-         * @returns {boolean | null} `true` when node is truthy, `false` when node is falsy,
-         *  `null` when it cannot be determined.
-         */
-        function getBooleanValue(node) {
-            if (node.value === null) {
-
-                /*
-                 * it might be a null literal or bigint/regex literal in unsupported environments .
-                 * https://github.com/estree/estree/blob/14df8a024956ea289bd55b9c2226a1d5b8a473ee/es5.md#regexpliteral
-                 * https://github.com/estree/estree/blob/14df8a024956ea289bd55b9c2226a1d5b8a473ee/es2020.md#bigintliteral
-                 */
-
-                if (node.raw === "null") {
-                    return false;
-                }
-
-                // regex is always truthy
-                if (typeof node.regex === "object") {
-                    return true;
-                }
-
-                return null;
-            }
-
-            return !!node.value;
-        }
-
-        /**
-         * Checks if a branch node of LogicalExpression short circuits the whole condition
-         * @param {ASTNode} node The branch of main condition which needs to be checked
-         * @param {string} operator The operator of the main LogicalExpression.
-         * @returns {boolean} true when condition short circuits whole condition
-         */
-        function isLogicalIdentity(node, operator) {
-            switch (node.type) {
-                case "Literal":
-                    return (operator === "||" && getBooleanValue(node) === true) ||
-                           (operator === "&&" && getBooleanValue(node) === false);
-
-                case "UnaryExpression":
-                    return (operator === "&&" && node.operator === "void");
-
-                case "LogicalExpression":
-
-                    /*
-                     * handles `a && false || b`
-                     * `false` is an identity element of `&&` but not `||`
-                     */
-                    return operator === node.operator &&
-                             (
-                                 isLogicalIdentity(node.left, operator) ||
-                                 isLogicalIdentity(node.right, operator)
-                             );
-
-                case "AssignmentExpression":
-                    return ["||=", "&&="].includes(node.operator) &&
-                        operator === node.operator.slice(0, -1) &&
-                        isLogicalIdentity(node.right, operator);
-
-                // no default
-            }
-            return false;
-        }
-
-        /**
-         * Checks if an identifier is a reference to a global variable.
-         * @param {ASTNode} node An identifier node to check.
-         * @returns {boolean} `true` if the identifier is a reference to a global variable.
-         */
-        function isReferenceToGlobalVariable(node) {
-            const scope = context.getScope();
-            const reference = scope.references.find(ref => ref.identifier === node);
-
-            return Boolean(
-                reference &&
-                reference.resolved &&
-                reference.resolved.scope.type === "global" &&
-                reference.resolved.defs.length === 0
-            );
-        }
-
-        /**
-         * Checks if a node has a constant truthiness value.
-         * @param {ASTNode} node The AST node to check.
-         * @param {boolean} inBooleanPosition `true` if checking the test of a
-         * condition. `false` in all other cases. When `false`, checks if -- for
-         * both string and number -- if coerced to that type, the value will
-         * be constant.
-         * @returns {Bool} true when node's truthiness is constant
-         * @private
-         */
-        function isConstant(node, inBooleanPosition) {
-
-            // node.elements can return null values in the case of sparse arrays ex. [,]
-            if (!node) {
-                return true;
-            }
-            switch (node.type) {
-                case "Literal":
-                case "ArrowFunctionExpression":
-                case "FunctionExpression":
-                    return true;
-                case "ClassExpression":
-                case "ObjectExpression":
-
-                    /**
-                     * In theory objects like:
-                     *
-                     * `{toString: () => a}`
-                     * `{valueOf: () => a}`
-                     *
-                     * Or a classes like:
-                     *
-                     * `class { static toString() { return a } }`
-                     * `class { static valueOf() { return a } }`
-                     *
-                     * Are not constant verifiably when `inBooleanPosition` is
-                     * false, but it's an edge case we've opted not to handle.
-                     */
-                    return true;
-                case "TemplateLiteral":
-                    return (inBooleanPosition && node.quasis.some(quasi => quasi.value.cooked.length)) ||
-                        node.expressions.every(exp => isConstant(exp, false));
-
-                case "ArrayExpression": {
-                    if (!inBooleanPosition) {
-                        return node.elements.every(element => isConstant(element, false));
-                    }
-                    return true;
-                }
-
-                case "UnaryExpression":
-                    if (
-                        node.operator === "void" ||
-                        node.operator === "typeof" && inBooleanPosition
-                    ) {
-                        return true;
-                    }
-
-                    if (node.operator === "!") {
-                        return isConstant(node.argument, true);
-                    }
-
-                    return isConstant(node.argument, false);
-
-                case "BinaryExpression":
-                    return isConstant(node.left, false) &&
-                            isConstant(node.right, false) &&
-                            node.operator !== "in";
-
-                case "LogicalExpression": {
-                    const isLeftConstant = isConstant(node.left, inBooleanPosition);
-                    const isRightConstant = isConstant(node.right, inBooleanPosition);
-                    const isLeftShortCircuit = (isLeftConstant && isLogicalIdentity(node.left, node.operator));
-                    const isRightShortCircuit = (inBooleanPosition && isRightConstant && isLogicalIdentity(node.right, node.operator));
-
-                    return (isLeftConstant && isRightConstant) ||
-                        isLeftShortCircuit ||
-                        isRightShortCircuit;
-                }
-                case "NewExpression":
-                    return inBooleanPosition;
-                case "AssignmentExpression":
-                    if (node.operator === "=") {
-                        return isConstant(node.right, inBooleanPosition);
-                    }
-
-                    if (["||=", "&&="].includes(node.operator) && inBooleanPosition) {
-                        return isLogicalIdentity(node.right, node.operator.slice(0, -1));
-                    }
-
-                    return false;
-
-                case "SequenceExpression":
-                    return isConstant(node.expressions[node.expressions.length - 1], inBooleanPosition);
-                case "SpreadElement":
-                    return isConstant(node.argument, inBooleanPosition);
-                case "CallExpression":
-                    if (node.callee.type === "Identifier" && node.callee.name === "Boolean") {
-                        if (node.arguments.length === 0 || isConstant(node.arguments[0], true)) {
-                            return isReferenceToGlobalVariable(node.callee);
-                        }
-                    }
-                    return false;
-                case "Identifier":
-                    return node.name === "undefined" && isReferenceToGlobalVariable(node);
-
-                // no default
-            }
-            return false;
-        }
-
-        /**
          * Tracks when the given node contains a constant condition.
          * @param {ASTNode} node The AST node to check.
          * @returns {void}
          * @private
          */
         function trackConstantConditionLoop(node) {
-            if (node.test && isConstant(node.test, true)) {
+            if (node.test && isConstant(context.getScope(), node.test, true)) {
                 loopsInCurrentScope.add(node);
             }
         }
@@ -280,7 +87,7 @@ module.exports = {
          * @private
          */
         function reportIfConstant(node) {
-            if (node.test && isConstant(node.test, true)) {
+            if (node.test && isConstant(context.getScope(), node.test, true)) {
                 context.report({ node: node.test, messageId: "unexpected" });
             }
         }
