+++ tests/lib/rules/no-constant-binary-expression.js
@@ -31,9 +31,6 @@ ruleTester.run("no-constant-binary-expression", rule, {
         "(x += 1) && foo",
         "`${bar}` && foo",
         "bar && foo",
-        "+1 && foo",
-        "-1 && foo",
-        "~1 && foo",
         "delete bar.baz && foo",
         "true ? foo : bar", // We leave ConditionalExpression for `no-constant-condition`.
         "new Foo() == true",
@@ -75,6 +72,9 @@ ruleTester.run("no-constant-binary-expression", rule, {
         { code: "true && foo", errors: [{ messageId: "constantShortCircuit" }] },
         { code: "'' && foo", errors: [{ messageId: "constantShortCircuit" }] },
         { code: "100 && foo", errors: [{ messageId: "constantShortCircuit" }] },
+        { code: "+100 && foo", errors: [{ messageId: "constantShortCircuit" }] },
+        { code: "-100 && foo", errors: [{ messageId: "constantShortCircuit" }] },
+        { code: "~100 && foo", errors: [{ messageId: "constantShortCircuit" }] },
         { code: "/[a-z]/ && foo", errors: [{ messageId: "constantShortCircuit" }] },
         { code: "Boolean([]) && foo", errors: [{ messageId: "constantShortCircuit" }] },
         { code: "Boolean() && foo", errors: [{ messageId: "constantShortCircuit" }] },
